[#84] ListRowView 분리 완료
